requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.0
PyYAML>=5.4.0
tqdm>=4.50.0
pytest>=7.0.0
scikit-learn>=0.24.0
pandas>=2.0.0
matplotlib>=3.3.0
tensorboard>=2.13.0
monai>=1.2.0 